Rename `XDRMatcher.h/cc` to `XDRQuery` to be more general.
